backend/requirements.txt
aiohttp>=3.9.3
aiosignal>=1.3.1
APScheduler>=3.10.4
asttokens>=2.4.1
attrs>=23.2.0
beautifulsoup4>=4.12.3
behave>=1.2.6
blinker>=1.7.0
Brotli>=1.1.0
cachetools>=5.3.3
certifi>=2024.2.2
cffi>=1.16.0
chardet>=5.2.0
charset-normalizer>=3.3.2
colorama>=0.4.6
comm>=0.2.0
cryptography>=42.0.5
cssselect>=1.2.0
debugpy>=1.8.0
decorator>=5.1.1
exceptiongroup>=1.2.0
execnet>=2.0.2
executing>=2.0.1
fasteners>=0.19
filelock>=3.13.1
frozenlist>=1.4.1
h11>=0.14.0
h2>=4.1.0
hpack>=4.0.0
hyperframe>=6.0.1
idna>=3.6
iniconfig>=2.0.0
ipinfo>=5.0.1
ipykernel>=6.27.1
ipython
jedi>=0.19.1
Jinja2>=3.1.3
jupyter_client>=8.6.0
jupyter_core>=5.5.0
kaitaistruct>=0.10
lxml>=5.1.0
markdown-it-py>=3.0.0
MarkupSafe>=2.1.5
matplotlib-inline>=0.1.6
mdurl>=0.1.2
multidict>=6.0.5
nest-asyncio>=1.5.8
numpy>=1.26.4
outcome>=1.3.0.post0
packaging>=23.2
pandas>=2.2.1
parameterized>=0.9.0
parse>=1.20.1
parse-type>=0.6.2
parso>=0.8.3
pdbp>=1.5.0
pdoc>=14.4.0
platformdirs>=4.2.0
pluggy>=1.4.0
prompt-toolkit>=3.0.41
psutil>=5.9.6
psycopg2-binary>=2.9.9
pure-eval>=0.2.2
py>=1.11.0
pyasn1>=0.5.1
pycparser>=2.21
pydivert>=2.1.0
Pygments>=2.17.2
pynose>=1.5.0
pyOpenSSL>=24.0.0
pyotp>=2.9.0
pyparsing>=3.1.1
pyreadline3>=3.4.1
PySocks>=1.7.1
pytest>=8.0.2
pytest-html>=2.0.1
pytest-metadata>=3.1.1
pytest-ordering>=0.6
pytest-rerunfailures>=13.0
pytest-xdist>=3.5.0
python-dateutil>=2.8.2
pytz>=2024.1
pywin32>=306
PyYAML>=6.0.1
pyzmq>=25.1.2
requests>=2.31.0
rich>=13.7.0
sbvirtualdisplay>=1.3.0
selenium
selenium-wire
seleniumbase
six>=1.16.0
sniffio>=1.3.1
sortedcontainers>=2.4.0
soupsieve>=2.5
stack-data>=0.6.3
tabcompleter>=1.3.0
tornado>=6.4
traitlets>=5.14.0
trio>=0.24.0
trio-websocket>=0.11.1
typing_extensions
tzdata>=2024.1
tzlocal>=5.2
urllib3>=2.2.1
wcwidth>=0.2.12
wsproto>=1.2.0
yarl>=1.9.4
zstandard>=0.22.0
